fix: adapt staged repair destinations to native index width Blocker-ID M43-STAGING-RECOVERY
Checkpoint: 5744ff6559df3c5ff027ae154a4f7867662129dd

Decision: cast only the host-repair destination view to int64 before the existing HiCache kernel.

Rationale: the native planner uses int32 destination metadata while HiCache requires int64; promotion and authoritative placement remain unchanged.

Validation: failed B200 staging-on run identified the dtype mismatch; AST and diff checks pass; rerun pending.

+++ python/sglang/srt/managers/hisparse_coordinator.py
@@ -862,7 +862,9 @@ class HiSparseCoordinator:
         repair_mask = valid_misses & ~promotion_mask
         promotion_dst = miss_dst[0][promotion_mask]
         repair_src = miss_src[0][repair_mask]
-        repair_dst = miss_dst[0][repair_mask]
+        # The placement planner stores destinations as int32, while the
+        # existing HiCache host-repair kernels require int64 indices.
+        repair_dst = miss_dst[0][repair_mask].to(dtype=torch.int64)
         if plan is not None:
             destination_tensor = self.mem_pool_device.kv_buffer[layer_id]
             destination_tensor[promotion_dst].copy_(plan.buffer[promotion_stage])
